// food-channel-processor/src/v1/channel-order-food/model/channel-food-revenue-sumary.output.data.model.ts
export class ChannelFoodRevenueSumaryOutputDataModel {

    total_order : number;

    total_order_SHF: number;
    
    total_order_GRF: number;

    total_order_GOF: number;

    total_order_BEF: number;

    total_revenue: number;

    total_revenue_SHF: number;

    total_revenue_GRF: number;

    total_revenue_GOF : number;

    total_revenue_BEF : number;

    percent_SHF: number;

    percent_GRF: number;

    percent_GOF : number;

    percent_BEF : number;

}